+++ scanpy/neighbors/_backends/gauss.py
@@ -1,10 +1,17 @@
+from __future__ import annotations
+
+from typing import TypeVar, Union
+
 import numpy as np
-from scipy.sparse import issparse
+from numpy.typing import NDArray
+from scipy.sparse import issparse, csr_matrix
 
 from .._common import _get_indices_distances_from_dense_matrix
 
 
-def _get_indices_distances_from_sparse_matrix(D, n_neighbors: int):
+def _get_indices_distances_from_sparse_matrix(
+    D: csr_matrix, n_neighbors: int
+) -> tuple[NDArray[np.int64], NDArray]:
     indices = np.zeros((D.shape[0], n_neighbors), dtype=int)
     distances = np.zeros((D.shape[0], n_neighbors), dtype=D.dtype)
     n_neighbors_m1 = n_neighbors - 1
@@ -28,10 +35,10 @@ def _get_indices_distances_from_sparse_matrix(D, n_neighbors: int):
 
 
 def _compute_connectivities_diffmap(
-    distances, n_neighbors, knn, density_normalize=True
+    distances: Union[np.ndarray, csr_matrix], n_neighbors: int
 ):
     # init distances
-    if knn:
+    if issparse(distances):
         Dsq = distances.power(2)
         indices, distances_sq = _get_indices_distances_from_sparse_matrix(
             Dsq, n_neighbors
@@ -48,7 +55,7 @@ def _compute_connectivities_diffmap(
 
     # choose sigma, the heuristic here doesn't seem to make much of a difference,
     # but is used to reproduce the figures of Haghverdi et al. (2016)
-    if knn:
+    if issparse(distances):
         # as the distances are not sorted
         # we have decay within the n_neighbors first neighbors
         sigmas_sq = np.median(distances_sq, axis=1)
@@ -64,21 +71,8 @@ def _compute_connectivities_diffmap(
         Den = np.add.outer(sigmas_sq, sigmas_sq)
         W = np.sqrt(Num / Den) * np.exp(-Dsq / Den)
         # make the weight matrix sparse
-        if not knn:
-            mask = W > 1e-14
-            W[~mask] = 0
-        else:
-            # restrict number of neighbors to ~k
-            # build a symmetric mask
-            mask = np.zeros(Dsq.shape, dtype=bool)
-            for i, row in enumerate(indices):
-                mask[i, row] = True
-                for j in row:
-                    if i not in set(indices[j]):
-                        W[j, i] = W[i, j]
-                        mask[j, i] = True
-            # set all entries that are not nearest neighbors to zero
-            W[~mask] = 0
+        mask = W > 1e-14
+        W[~mask] = 0
     else:
         W = Dsq.copy()  # need to copy the distance matrix here; what follows is inplace
         for i in range(len(Dsq.indptr[:-1])):
